figure-caption-extrac/src/config/config_loader.py
from pathlib import Path
import yaml
from src.models.AppConfig import AppConfig

CONFIG_PATH = Path(__file__).parent / "config.yaml"

def load_config(path: Path = CONFIG_PATH) -> AppConfig:
    with open(path, "r") as f:
        raw = yaml.safe_load(f)
    return AppConfig(**raw)