figure-caption-extrac/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.13"
anyio==4.9.0 ; python_version >= "3.13"
certifi==2025.4.26 ; python_version >= "3.13"
charset-normalizer==3.4.2 ; python_version >= "3.13"
click==8.1.8 ; python_version >= "3.13"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.13"
coverage==7.8.0 ; python_version >= "3.13"
duckdb==1.2.2 ; python_version >= "3.13"
fastapi==0.115.12 ; python_version >= "3.13"
h11==0.16.0 ; python_version >= "3.13"
httpcore==1.0.9 ; python_version >= "3.13"
httpx==0.28.1 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
iniconfig==2.1.0 ; python_version >= "3.13"
markdown-it-py==3.0.0 ; python_version >= "3.13"
mdurl==0.1.2 ; python_version >= "3.13"
numpy==2.2.6 ; python_version >= "3.13"
packaging==25.0 ; python_version >= "3.13"
pandas==2.2.3 ; python_version >= "3.13"
pluggy==1.6.0 ; python_version >= "3.13"
pydantic-core==2.33.2 ; python_version >= "3.13"
pydantic==2.11.4 ; python_version >= "3.13"
pygments==2.19.1 ; python_version >= "3.13"
pytest==8.3.5 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
python-dotenv==1.1.0 ; python_version >= "3.13"
pytz==2025.2 ; python_version >= "3.13"
pyyaml==6.0.2 ; python_version >= "3.13"
requests==2.32.3 ; python_version >= "3.13"
rich==14.0.0 ; python_version >= "3.13"
shellingham==1.5.4 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
sniffio==1.3.1 ; python_version >= "3.13"
starlette==0.46.2 ; python_version >= "3.13"
typer==0.15.4 ; python_version >= "3.13"
typing-extensions==4.13.2 ; python_version >= "3.13"
typing-inspection==0.4.0 ; python_version >= "3.13"
tzdata==2025.2 ; python_version >= "3.13"
urllib3==2.4.0 ; python_version >= "3.13"
uvicorn==0.34.2 ; python_version >= "3.13"
watchdog==6.0.0 ; python_version >= "3.13"
